feat(coprocessor): bigint chain ID across components
Use a bigint for chain ID in the coprocessor, meaning 63 bits of useful
data can be used (instead of 31 bits in some cases).

Using i64 instead of u64 due to more simple mapping to bigint in the DB.

+++ coprocessor/fhevm-engine/tfhe-worker/src/db_queries.rs
@@ -59,7 +59,7 @@ pub async fn check_if_api_key_is_valid<T>(
 }
 
 pub struct FetchTenantKeyResult {
-    pub chain_id: i32,
+    pub chain_id: i64,
     pub verifying_contract_address: String,
     pub acl_contract_address: String,
     pub server_key: tfhe::ServerKey,
